+++ src/ibmdiagrams/ibmbase/types.py
@@ -559,7 +559,8 @@ class Types:
       return data
 
    def buildDrawioShape(self, id, node, x, y, width, height, meta, genflag):
-      labelsize = 20
+      #labelsize = 20
+      labelsize = 15
       shape = node["shape"].lower()
       name = node["label"]
       subname = node["sublabel"]